first commit 11

+++ resources/views/participation/index.blade.php
@@ -37,7 +37,7 @@
                 <form method="POST" action="{{ route('search.participation') }}">
                     @csrf
                     <div class="row">
-                        <div class="col-lg-2">
+                        <div class="col-lg-3">
                             <label>Heure</label>
                             <select class="form-control" id="heure_id" name="heure_id" required="">
                                 <option value="">Selectionner</option>
@@ -47,7 +47,7 @@
 
                             </select>
                         </div>
-                        {{--<div class="col-3">
+                       <div class="col-3">
                             <label>Département</label>
                             <select class="form-control" id="departement_id" name="departement_id" >
                                 <option value="">Selectionner</option>
@@ -57,45 +57,14 @@
                                 @endforeach
                             </select>
                         </div>
-                        <div class="col-3">
-                            <label>Arrondissement</label>
-                            <select class="form-control" id="arrondissement_id" name="arrondissement_id">
-                                <option value="">Veuillez selectionner </option>
-                                @foreach ($arrondissements as $arrondissement)
-                                <option value="{{$arrondissement->id}}" {{ $arrondissement_id==$arrondissement->id ? 'selected' : '' }}>{{$arrondissement->nom}}</option>
-                                    @endforeach
+                        <div class="col-lg-3">
+                            <label>Heure</label>
+                            <select class="form-control" id="etat" name="etat"  >
+                                <option value="">Selectionner</option>
+                                <option value="reseigner" {{$etat=="reseigner" ? 'selected' : ''}}>Renseigner</option>
+                                <option value="nonrenseigner" {{$etat=="nonrenseigner" ? 'selected' : ''}}>Non Renseigner</option>
                             </select>
                         </div>
-                              <div class="col-2">
-                                <label>Commune</label>
-                                <select class="form-control" id="commune_id" name="commune_id" >
-                                    <option value="">Veuillez selectionner </option>
-        
-                                    @foreach ($communes as $commune)
-                                    <option value="{{$commune->id}}" {{ $commune_id==$commune->id ? 'selected' : '' }}>{{$commune->nom}}</option>
-                                        @endforeach
-                                </select>
-                            </div>
-                                <div class="col-3">
-                                    <label>centrevote</label>
-                                    <select class="form-control" name="centrevote_id" id="centrevote_id" >
-                                        <option value="">Veuillez selectionner </option>
-        
-                                        @foreach ($centreVotes as $centreVote)
-                                        <option value="{{$centreVote->id}}" {{ $centrevote_id==$centreVote->id ? 'selected' : '' }}>{{$centreVote->nom}}</option>
-                                            @endforeach
-                                    </select>
-                                </div>
-                                <div class="col-2">
-                                    <label>Bureau  de vote</label>
-                                    <select class="form-control" name="lieuvote_id" id="lieuvote_id" >
-                                        <option value="">Veuillez selectionner </option>
-        
-                                        @foreach ($lieuVotes as $lieuVote)
-                                        <option value="{{$lieuVote->id}}" {{ $lieuvote_id==$lieuVote->id ? 'selected' : '' }}>{{$lieuVote->nom}}</option>
-                                            @endforeach
-                                    </select>
-                                </div> --}}
         
     
                         <div class="col-lg-2">
@@ -107,7 +76,7 @@
                 <table id="datatable-buttons" class="table table-bordered table-responsive-md table-striped text-center">
                     <thead>
                         <tr>
-                            <th>#</th>
+                           
                           {{--   <th>Region</th> --}}
                             <th>Departement</th>
                             <th>Heure</th>
@@ -122,22 +91,21 @@
                     <tbody>
                     @foreach ($participations as $participation)
                         <tr>
-                            <td>{{ $participation->id }}</td>
                          {{--    <td>{{ $participation->region->nom }}</td> --}}
                             
-                            <td> @if ($participation->departement)
+                            <td> @if (!empty($participation->departement))
                                 
-                               {{ $participation->departement->nom }} @endif</td>
-                            <td>{{ $participation->heure->designation }}</td>
+                               {{ $participation->departement }} @endif</td>
+                            <td>@if (!empty($participation->heure)){{ $participation->heure }} @endif</td>
                             
-                            <td>{{ $participation->resultat }}</td>
-                            <td>{{ $participation->lieuvote->centrevote->nom }}</td>
-                            <td>{{ $participation->lieuvote->nom }}</td>
+                            <td>@if (!empty($participation->resultat)){{ $participation->resultat }} @endif</td>
+                            <td>{{ $participation->centrevote }}</td>
+                            <td>{{ $participation->lieuvote  }}</td>
                              <td>
                                {{--  <a href="{{ route('participation.edit', $participation->id) }}" role="button" class="btn btn-primary"><i class="fas fa-edit"></i></a> --}}
-                                {!! Form::open(['method' => 'DELETE', 'route'=>['participation.destroy', $participation->id], 'style'=> 'display:inline', 'onclick'=>"if(!confirm('Êtes-vous sûr de vouloir supprimer cet enregistrement ?')) { return false; }"]) !!}
+                               @if (!empty($participation->id)) {!! Form::open(['method' => 'DELETE', 'route'=>['participation.destroy', $participation->id], 'style'=> 'display:inline', 'onclick'=>"if(!confirm('Êtes-vous sûr de vouloir supprimer cet enregistrement ?')) { return false; }"]) !!}
                                 <button class="btn btn-danger"><i class="far fa-trash-alt"></i></button>
-                                {!! Form::close() !!}
+                                {!! Form::close() !!} @endif
 
 
 
@@ -156,145 +124,4 @@
     </div>
 </div>
 
-@endsection
-@section('script')
-<script>
-
-    url_app = '{{ config('app.url') }}';
- /* $("#region_id").change(function () {
-    var region_id =  $("#region_id").children("option:selected").val();
-    $(".region").val(region_id);
-    $(".departement").val("");
-    $(".commune").val("");
-        var departement = "<option value=''>Veuillez selectionner</option>";
-        $.ajax({
-            type:'GET',
-            url:url_app+'/departement/by/region/'+region_id,
-        //   url:'http://vmi435145.contaboserver.net:9000/departement/by/region/'+region_id,
-          // url:'http://127.0.0.1/gestionmateriel/public/departement/by/region/'+region_id,
-          //    url:'http://127.0.0.1:8000/departement/by/region/'+region_id,
-            data:'_token = <?php echo csrf_token() ?>',
-            success:function(data) {
-
-                $.each(data,function(index,row){
-                    //alert(row.nomd);
-                    departement +="<option value="+row.id+">"+row.nom+"</option>";
-
-                });
-                $("#departement_id").empty();
-                $("#commune_id").empty();
-                $("#departement_id").append(departement);
-            }
-        });
-    });*/
-    $("#departement_id").change(function () {
-        var departement_id =  $("#departement_id").children("option:selected").val();
-        $(".departement").val(departement_id);
-        $(".commune").val("");
-        $("#commune_id").empty();
-        $("#arrondissement_id").empty();
-        $("#centrevote_id").empty();
-        $("#lieuvote_id").empty();
-            var arrondissement = "<option value=''>Veuillez selectionner</option>";
-            $.ajax({
-                type:'GET',
-                url:url_app+'/arrondissement/by/departement/'+departement_id,
-                data:'_token = <?php echo csrf_token() ?>',
-                success:function(data) {
-
-                    $.each(data,function(index,row){
-                        //alert(row.nomd);
-                        arrondissement +="<option value="+row.id+">"+row.nom+"</option>";
-
-                    });
-                    $("#arrondissement_id").empty();
-                    $("#arrondissement_id").append(arrondissement);
-                }
-            });
-        });
-    $("#arrondissement_id").change(function () {
-        var arrondissement_id =  $("#arrondissement_id").children("option:selected").val();
-        $(".commune").val("");
-        $("#commune_id").empty();
-        $("#centrevote_id").empty();
-        $("#lieuvote_id").empty();
-            var commune = "<option value=''>Veuillez selectionner</option>";
-            $.ajax({
-                type:'GET',
-                url:url_app+'/commune/by/arrondissement/'+arrondissement_id,
-                 data:'_token = <?php echo csrf_token() ?>',
-                success:function(data) {
-
-                    $.each(data,function(index,row){
-                        //alert(row.nomd);
-                        commune +="<option value="+row.id+">"+row.nom+"</option>";
-
-                    });
-                    $("#commune_id").empty();
-                    $("#commune_id").append(commune);
-                }
-            });
-        });
-        $("#commune_id").change(function () {
-            var commune_id =  $("#commune_id").children("option:selected").val();
-                var centrevote = "<option value=''>Veuillez selectionner</option>";
-                $("#centrevote_id").empty();
-                $("#lieuvote_id").empty();
-                $.ajax({
-                    type:'GET',
-                    url:url_app+'/centrevote/by/commune/'+commune_id,
-
-                    data:'_token = <?php echo csrf_token() ?>',
-                    success:function(data) {
-
-                        $.each(data,function(index,row){
-                            //alert(row.nomd);
-                            centrevote +="<option value="+row.id+">"+row.nom+"</option>";
-
-                        });
-
-                        $("#centrevote_id").append(centrevote);
-                    }
-                });
-            });
-
-        $("#centrevote_id").change(function () {
-                var centrevote_id =  $("#centrevote_id").children("option:selected").val();
-                    var lieuvote = "<option value=''>Veuillez selectionner</option>";
-                    $("#lieuvote_id").empty();
-                    $.ajax({
-                        type:'GET',
-                        url:url_app+'/lieuvote/temoin/participation/by/centrevote/'+centrevote_id,
-                        data:'_token = <?php echo csrf_token() ?>',
-                        success:function(data) {
-
-                            $.each(data,function(index,row){
-                              //  alert(row.id);
-                                lieuvote +="<option value="+row.id+">"+row.nom+"</option>";
-
-                            });
-
-                            $("#lieuvote_id").append(lieuvote);
-                        }
-                    });
-                });
-                $("#lieuvote_id").change(function () {
-                var lieuvote_id =  $("#lieuvote_id").children("option:selected").val();
-                    $.ajax({
-                        type:'GET',
-                        url:url_app+'/electeur/by/lieuvote/'+lieuvote_id,
-                        data:'_token = <?php echo csrf_token() ?>',
-                        success:function(data) {
-                         //   alert(data)
-
-                            $('#electeur').empty()
-                           $('#electeur').append("<h4> Nombre Electeurs : "+data.nb+"</h4>")
-                           $('#nb_electeur').val(data.nb)
-
-                        }
-                    });
-                });
-              
-
-</script>
 @endsection
